// src/api/models/business-service-request-type-edition-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessServiceRequestTypeVisibility } from './business-service-request-type-visibility';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestStatus } from './service-request-status';
// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface BusinessServiceRequestTypeEditionModel
 */
export interface BusinessServiceRequestTypeEditionModel {
    /**
     * 
     * @type {Array<string>}
     * @memberof BusinessServiceRequestTypeEditionModel
     */
    'sendNotificationsToEmailList'?: Array<string> | null;
    /**
     * 
     * @type {Array<ServiceRequestStatus>}
     * @memberof BusinessServiceRequestTypeEditionModel
     */
    'sendNotificationsByRequestStatusList'?: Array<ServiceRequestStatus> | null;
    /**
     * 
     * @type {BusinessServiceRequestTypeVisibility}
     * @memberof BusinessServiceRequestTypeEditionModel
     */
    'visibility'?: BusinessServiceRequestTypeVisibility;
    /**
     * 
     * @type {number}
     * @memberof BusinessServiceRequestTypeEditionModel
     */
    'id'?: number;
    /**
     * 
     * @type {Status}
     * @memberof BusinessServiceRequestTypeEditionModel
     */
    'status'?: Status;
}