// src/api/models/business-service-request-type-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { BusinessModel } from './business-model';
// May contain unused imports in some cases
// @ts-ignore
import { BusinessServiceRequestTypeVisibility } from './business-service-request-type-visibility';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestStatus } from './service-request-status';
// May contain unused imports in some cases
// @ts-ignore
import { ServiceRequestTypeModel } from './service-request-type-model';
// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface BusinessServiceRequestTypeModel
 */
export interface BusinessServiceRequestTypeModel {
    /**
     * 
     * @type {number}
     * @memberof BusinessServiceRequestTypeModel
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof BusinessServiceRequestTypeModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {Status}
     * @memberof BusinessServiceRequestTypeModel
     */
    'status'?: Status;
    /**
     * 
     * @type {string}
     * @memberof BusinessServiceRequestTypeModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessServiceRequestTypeModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof BusinessServiceRequestTypeModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BusinessServiceRequestTypeModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BusinessServiceRequestTypeModel
     */
    'requestTypeId'?: number;
    /**
     * 
     * @type {ServiceRequestTypeModel}
     * @memberof BusinessServiceRequestTypeModel
     */
    'requestType'?: ServiceRequestTypeModel;
    /**
     * 
     * @type {number}
     * @memberof BusinessServiceRequestTypeModel
     */
    'businessId'?: number;
    /**
     * 
     * @type {BusinessModel}
     * @memberof BusinessServiceRequestTypeModel
     */
    'business'?: BusinessModel;
    /**
     * 
     * @type {BusinessServiceRequestTypeVisibility}
     * @memberof BusinessServiceRequestTypeModel
     */
    'visibility'?: BusinessServiceRequestTypeVisibility;
    /**
     * 
     * @type {Array<string>}
     * @memberof BusinessServiceRequestTypeModel
     */
    'sendNotificationsToEmailList'?: Array<string> | null;
    /**
     * 
     * @type {Array<ServiceRequestStatus>}
     * @memberof BusinessServiceRequestTypeModel
     */
    'sendNotificationsByRequestStatusList'?: Array<ServiceRequestStatus> | null;
}