// src/api/models/service-request-type-creation-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ServiceRequestTypeCreationModel
 */
export interface ServiceRequestTypeCreationModel {
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeCreationModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeCreationModel
     */
    'code'?: string | null;
}