// src/api/models/service-request-type-edition-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface ServiceRequestTypeEditionModel
 */
export interface ServiceRequestTypeEditionModel {
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeEditionModel
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeEditionModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestTypeEditionModel
     */
    'id'?: number;
    /**
     * 
     * @type {Status}
     * @memberof ServiceRequestTypeEditionModel
     */
    'status'?: Status;
}