// src/api/models/service-request-type-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Argento Oficina Virtual
 * Api para Portal de Clientes Argento
 *
 * The version of the OpenAPI document: 0.10.2
 * Contact: info@mts.com.do
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Status } from './status';

/**
 * 
 * @export
 * @interface ServiceRequestTypeModel
 */
export interface ServiceRequestTypeModel {
    /**
     * 
     * @type {number}
     * @memberof ServiceRequestTypeModel
     */
    'id'?: number;
    /**
     * 
     * @type {boolean}
     * @memberof ServiceRequestTypeModel
     */
    'isNullable'?: boolean;
    /**
     * 
     * @type {Status}
     * @memberof ServiceRequestTypeModel
     */
    'status'?: Status;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeModel
     */
    'createdBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeModel
     */
    'createdOnUtc'?: string;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeModel
     */
    'updatedBy'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeModel
     */
    'updatedOnUtc'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeModel
     */
    'code'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ServiceRequestTypeModel
     */
    'name'?: string | null;
}